added basic barrel heat control (#61)

+++ src/bsp/inc/bsp_uart.h
@@ -16,6 +16,6 @@ typedef struct _UART_Instance
 } UART_Instance_t;
 
 void UART_Service_Init(UART_Instance_t *uart_insatce);
-UART_Instance_t *UART_Register(UART_HandleTypeDef *huart, uint8_t *rx_buffer, uint8_t rx_buffer_size, void (*callback)(UART_Instance_t *uart_instance)); 
+UART_Instance_t *UART_Register(UART_HandleTypeDef *huart, uint8_t *rx_buffer, uint16_t rx_buffer_size, void (*callback)(UART_Instance_t *uart_instance)); 
 HAL_StatusTypeDef UART_Transmit(UART_Instance_t *uart_instance, uint8_t *tx_buffer, uint16_t tx_buffer_size, uint8_t send_type);
 #endif // BSP_UART_H
